requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.6
python-dotenv>=1.1.1
requests>=2.32.4
pydantic>=2.8.2
